refactor: move cloud provider mapping to its own package
