+++ pkg/config/v0/oci_oke_kubernetes_runtime_instance.go
@@ -8,10 +8,10 @@ import (
 	"net/http"
 	"time"
 
-	apilib "github.com/threeport/threeport/pkg/api/lib/v0"
 	api_v0 "github.com/threeport/threeport/pkg/api/v0"
 	client_lib "github.com/threeport/threeport/pkg/client/lib/v0"
 	client_v0 "github.com/threeport/threeport/pkg/client/v0"
+	mapping "github.com/threeport/threeport/pkg/mapping/v0"
 	util "github.com/threeport/threeport/pkg/util/v0"
 )
 
@@ -176,7 +176,7 @@ func (o *OciOkeKubernetesRuntimeInstanceConfig) Create(
 	}
 
 	// get location for OCI region
-	location, err := apilib.GetLocationForOciRegion(*ociOkeKubernetesRuntimeInstanceValues.Region)
+	location, err := mapping.GetLocationForOciRegion(*ociOkeKubernetesRuntimeInstanceValues.Region)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get Threeport location for OCI region %s: %w", *ociOkeKubernetesRuntimeInstanceValues.Region, err)
 	}
