+++ pkg/mapping/v0/location_test.go
